refactor: move fetchWithTimeout to fetch-client
